Linted the test contracts

+++ contracts/testing/TestOracle.sol
@@ -8,23 +8,27 @@ pragma solidity ^0.8.9;
 // not actually inheriting it because I dont want to implement the whole thing.abi
 
 contract TestOracle {
-    uint price;
-    constructor(uint _price) {
+    uint256 price;
+
+    constructor(uint256 _price) {
         price = _price;
     }
 
-    function latestRoundData() external
-    view returns (
-      uint80 roundId,
-      int256 answer,
-      uint256 startedAt,
-      uint256 updatedAt,
-      uint80 answeredInRound
-    ) {
+    function latestRoundData()
+        external
+        view
+        returns (
+            uint80 roundId,
+            int256 answer,
+            uint256 startedAt,
+            uint256 updatedAt,
+            uint80 answeredInRound
+        )
+    {
         return (uint80(1), int256(price), uint256(1), uint256(1), uint80(1));
     }
-    
-    function setPrice(uint _price) public {
+
+    function setPrice(uint256 _price) public {
         price = _price;
     }
 }
